Move Postgres LineageResource definition to it's own source file.
